@@ -1,59 +1,74 @@
 <script>
-	import Counter from './Counter.svelte';
-	import welcome from '$lib/images/svelte-welcome.webp';
-	import welcome_fallback from '$lib/images/svelte-welcome.png';
+	let start = $state(2);
+	let kids = $state(2);
+	let birthAge = $state(20);
+	let generations = $state(1);
+	let populationPerGeneration = $derived.by(() => {
+		const result = [{
+			population: start,
+			currentGenerationPopulation: start
+		}];
+		if (generations === 1) return result;
+		
+		for(let i = 1; i < generations; i++) {
+			const currentGenerationPopulation = result[i-1].currentGenerationPopulation * kids;
+			console.log(result[i-1].population);
+			const population = result[i-1].population + currentGenerationPopulation;
+			result.push({ population, currentGenerationPopulation });
+		}
+		return result;
+	});
+	
+	$effect(() => {
+		console.log('populationPerGeneration changed', populationPerGeneration.length);
+	})
 </script>
 
-<svelte:head>
-	<title>Home</title>
-	<meta name="description" content="Svelte demo app" />
-</svelte:head>
+<div class="max-w-3xl mx-auto">
+<div class="mt-10 grid grid-cols-1 gap-6 sm:grid-cols-3 p-6">
+	<div>
+		<label for="first-name" class="block text-sm/6 font-medium">KIDS (per adult)</label>
+		<div class="mt-2">
+			<input type="number" bind:value={kids} class="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm/6">
+		</div>
+	</div>
+	<div>
+		<label for="last-name" class="block text-sm/6 font-medium">GENERATIONS</label>
+		<div class="mt-2">
+			<input type="number" bind:value={generations} class="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm/6">
+		</div>
+	</div>
+	<div>
+		<label for="last-name" class="block text-sm/6 font-medium">AGE (when kids are born)</label>
+		<div class="mt-2">
+			<input type="number" bind:value={birthAge} class="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm/6">
+		</div>
+	</div>
+</div>
 
-<section>
-	<h1>
-		<span class="welcome">
-			<picture>
-				<source srcset={welcome} type="image/webp" />
-				<img src={welcome_fallback} alt="Welcome" />
-			</picture>
-		</span>
+<div class="tablewrapper text-center mt-8 w-full" style="max-height: 60vh; overflow: auto;">
+	<table class="w-full">
+		<tbody>
+			<tr class="sticky top-0" style="background: #fff">
+				<th class="p-2">Generation</th>
+				<th class="p-2">Actual population size</th>
+				<th class="p-2">Total population</th>
+			</tr>
+			{#each populationPerGeneration.reverse() as population, i}
+				<tr>
+					<td class="p-2">{i+1}</td>
+					<td class="p-2">{population.currentGenerationPopulation.toLocaleString('de-DE', { maximumFractionDigits: 0 })}</td>
+					<td class="p-2">{population.population.toLocaleString('de-DE', { maximumFractionDigits: 0 })}</td>
+				</tr>
+			{/each}
+		</tbody>
+	</table>
+</div>
 
-		to your new<br />SvelteKit app
-	</h1>
-
-	<h2>
-		try editing <strong>src/routes/+page.svelte</strong>
-	</h2>
-
-	<Counter />
-</section>
-
+<p class="w-full text-3xl text-center font-bold mt-8">{populationPerGeneration[populationPerGeneration.length-1].population.toLocaleString('de-DE', { maximumFractionDigits: 0 })} humans in {(generations - 1)*birthAge} years</p>
+</div>
 <style>
-	section {
-		display: flex;
-		flex-direction: column;
-		justify-content: center;
-		align-items: center;
-		flex: 0.6;
-	}
-
-	h1 {
-		width: 100%;
-	}
-
-	.welcome {
-		display: block;
-		position: relative;
-		width: 100%;
-		height: 0;
-		padding: 0 0 calc(100% * 495 / 2048) 0;
-	}
-
-	.welcome img {
-		position: absolute;
-		width: 100%;
-		height: 100%;
-		top: 0;
-		display: block;
+	:global(body) {
+		background: #fff;
 	}
 </style>
